pyproject.toml
[tool.poetry]
name = "tweet_tips_predictor"
version = "0.1.0"
description = "Small API for predicting if a text of a tweet is a Betting tip or no."
authors = ["José Victor <josevictorferreira@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.10.3"
flask = "^1.0"
pymongo = "^3.8"
pytest-mongodb = "^2.1"
pandas = "^0.24.2"
sklearn = "^0.0.0"

[tool.poetry.dev-dependencies]
ipython = "^7.6"
pylint = "^2.3"
pytest = "^5.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
